ci: Include files
