+++ .npmignore
@@ -1,4 +1,5 @@
 node_modules
+coverage
 src
 *.config.js
 .*
