+++ package.json
@@ -20,49 +20,45 @@
       "url": "https://cutterscrossing.com"
     }
   ],
-  "main": "./dist/moment-extension.js",
-  "module": "./dist/moment-extension.js",
+  "files": [
+    "dist"
+  ],
+  "type": "module",
+  "main": "./dist/moment-extension.umd.cjs",
+  "module": "./dist/moment-extension.mjs",
+  "exports": {
+    ".": {
+      "import": "./dist/moment-extension.mjs",
+      "require": "./dist/moment-extension.umd.cjs"
+    }
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/cutterbl/moment-extension.git"
   },
   "scripts": {
-    "build": "rollup -c ./.configs/rollup.config.mjs",
-    "coverage": "is-ci || jest --coverage",
-    "prepare": "is-ci || husky install",
-    "release": "standard-version -a",
-    "shrinkwrap": "zx ./.scripts/shrinkwrap.mjs",
-    "test": "jest"
+    "prepare": "node .husky/install.mjs",
+    "build": "vite build",
+    "test": "vitest",
+    "coverage": "vitest run --coverage",
+    "release": "commit-and-tag-version",
+    "shrinkwrap": "zx ./.scripts/shrinkwrap.mjs"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.3",
-    "@babel/eslint-parser": "^7.23.3",
-    "@babel/plugin-transform-runtime": "^7.23.4",
-    "@babel/preset-env": "^7.23.3",
-    "@commitlint/cli": "^18.4.3",
-    "@commitlint/config-conventional": "^18.4.3",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-eslint": "^9.0.5",
-    "@rollup/plugin-terser": "0.4.4",
-    "babel-jest": "^27.5.1",
-    "core-js": "^3.33.3",
-    "eslint": "^8.54.0",
-    "eslint-config-prettier": "^9.0.0",
-    "expect": "^27.5.1",
-    "husky": "^7.0.4",
+    "@commitlint/cli": "^18.6.1",
+    "@commitlint/config-conventional": "^18.6.3",
+    "@vitest/coverage-v8": "^1.6.0",
+    "commit-and-tag-version": "^12.4.1",
+    "husky": "^9.0.11",
     "is-ci": "^3.0.1",
-    "jest": "^27.5.1",
-    "prettier": "^3.1.0",
-    "regenerator-runtime": "^0.14.0",
-    "rollup": "^4.6.0",
-    "rollup-plugin-cleanup": "^3.2.1",
-    "rollup-plugin-clear": "^2.0.7",
-    "standard-version": "^9.5.0",
-    "zx": "^7.2.3"
+    "prettier": "^3.3.2",
+    "terser": "^5.31.1",
+    "vite": "^5.3.1",
+    "vitest": "^1.6.0",
+    "zx": "^8.1.2"
   },
   "dependencies": {
-    "@babel/runtime-corejs3": "^7.23.4",
-    "moment": "^2.29.4",
-    "moment-timezone": "^0.5.43"
+    "moment": "^2.30.1",
+    "moment-timezone": "^0.5.45"
   }
 }
